django-backend/requirements/macos-requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.3.1
Django==3.2.22
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
factory-boy==3.3.0
Faker==19.6.2
flake8==6.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
mccabe==0.7.0
multidict==6.0.4
packaging==23.1
pamqp==3.2.1
pika==1.3.2
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycodestyle==2.11.0
pyflakes==3.1.0
pytest==7.4.2
pytest-django==4.5.2
pytest-env==1.0.1
pytest-mock==3.11.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
google-auth==2.23.3
django-storages==1.14.2
boto3==1.28.65
PyJWT==2.8.0
django-nonrelated-inlines==0.2
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.11.2
freezegun==1.2.2
openpyxl==3.1.2
Pillow==10.1.0
novita-client==0.4.12
stripe==7.11.
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==8.0
sentry-sdk==1.40.0
moviepy==1.0.3
django-debug-toolbar==4.3.0